# lib/api-zod/src/generated/types/createPropertyBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Clinic Launch OS API
 * OpenAPI spec version: 0.1.0
 */
import type { CreatePropertyBodyStatus } from "./createPropertyBodyStatus";

export interface CreatePropertyBody {
  address?: string | null;
  postcode?: string | null;
  sqFootage?: number | null;
  annualRentGbp?: number | null;
  monthlyRentGbp?: number | null;
  vatOnRent?: boolean | null;
  businessRatesGbp?: number | null;
  serviceChargeGbp?: number | null;
  leaseLength?: string | null;
  useClass?: string | null;
  availabilityDate?: string | null;
  parkingSpaces?: number | null;
  frontageMeters?: number | null;
  agentName?: string | null;
  agentPhone?: string | null;
  agentEmail?: string | null;
  status?: CreatePropertyBodyStatus;
  notes?: string | null;
}