# lib/api-zod/src/generated/types/launchPhase.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Clinic Launch OS API
 * OpenAPI spec version: 0.1.0
 */
import type { LaunchPhaseStatus } from "./launchPhaseStatus";

export interface LaunchPhase {
  id: number;
  projectId: number;
  name: string;
  description?: string | null;
  sortOrder: number;
  status: LaunchPhaseStatus;
  totalCostLow: number;
  totalCostMid: number;
  totalCostHigh: number;
  taskCount: number;
  completedTaskCount: number;
  createdAt: string;
  updatedAt: string;
}