# lib/api-zod/src/generated/types/updateTaskBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Clinic Launch OS API
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateTaskBodyCostTier } from "./updateTaskBodyCostTier";
import type { UpdateTaskBodyRiskLevel } from "./updateTaskBodyRiskLevel";
import type { UpdateTaskBodyStatus } from "./updateTaskBodyStatus";

export interface UpdateTaskBody {
  title?: string;
  description?: string | null;
  owner?: string | null;
  contractor?: string | null;
  supplier?: string | null;
  status?: UpdateTaskBodyStatus;
  riskLevel?: UpdateTaskBodyRiskLevel;
  costTier?: UpdateTaskBodyCostTier;
  costLow?: number;
  costMid?: number;
  costHigh?: number;
  dueDate?: string | null;
  durationDays?: number | null;
  dependencies?: number[] | null;
  notes?: string | null;
  isNonNegotiable?: boolean;
  isCriticalRisk?: boolean;
  sortOrder?: number;
}